requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
databricks-cli==0.16.4
idna==3.3
iniconfig==1.1.1
joblib==1.1.0
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
numpy==1.23.0
oauthlib==3.2.0
packaging==21.0
pandas==1.4.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==0.13.1
protobuf==4.21.2
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
tabulate==0.8.9
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
